feat: Configurations for API, blank, but could be useful idea
A file that should contain API end points, their textual representations, with possible options. Since the API is very new, it can be possible that it is fixed or changed slightly in the future. To avoid breaking other components, we can isolate the api endpoints, their actual textual represnetations, and not have to worry about the change everywhere else in the code

Need to be discussed in detail with team