feat: Add basic API interfacing with APIv4 @ Mapillary
This is meant to implement logically, as code, the API interface with which to access the APIv4 web interfaces. I thought the URLs itself did not seem intuitive at first, so I made some wrapper functions around them to promote compatibility that is, if the URLs were to change tomorrow for some reason, we can simply fix that adding editing a single line, and breaking nothing in the code base.

Changing the configurations here in the functions should be enough to accompany what the API end points are. Maybe even adding new functions or types of end points.