feat: Make basic test ready to use
Need to go through each of the use cases and functionalities to make sure the tests align properly. This is just a template for what a test can look like. Please refer to the requests library for understanding how pytests integrate with the project, since it's more or less almost the same approach